requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.8
numpy>=1.23
Pillow>=10.0.0
tensorboard>=2.14.0
scikit-learn>=1.2.0
joblib>=1.3.0

